requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg2-binary
dotenv
dj-database-url
supabase